LayoutTests/platform/ios-simulator/editing/inserting/insert-paste-bidi-control-expected.txt

a‫b!‬
‫b!‬
a‫b!‬
‫b!‬
==================================
FAILED: copy/paste a&#x202B;b!&#x202C;
actual (character in Unicode value): 60 102 111 110 116 32 115 105 122 101 61 34 51 34 62 60 115 112 97 110 32 115 116 121 108 101 61 34 45 119 101 98 107 105 116 45 116 101 120 116 45 115 105 122 101 45 97 100 106 117 115 116 58 32 97 117 116 111 59 32 98 97 99 107 103 114 111 117 110 100 45 99 111 108 111 114 58 32 114 103 98 97 40 50 53 53 44 32 50 53 53 44 32 50 53 53 44 32 48 41 59 34 62 97 8235 98 33 8236 60 47 115 112 97 110 62 60 47 102 111 110 116 62
expected (character in Unicode value): 97 8235 98 33 8236
==================================
FAILED: copy/paste &#x202B;b!&#x202C;
actual (character in Unicode value): 60 102 111 110 116 32 115 105 122 101 61 34 51 34 62 60 115 112 97 110 32 115 116 121 108 101 61 34 45 119 101 98 107 105 116 45 116 101 120 116 45 115 105 122 101 45 97 100 106 117 115 116 58 32 97 117 116 111 59 32 98 97 99 107 103 114 111 117 110 100 45 99 111 108 111 114 58 32 114 103 98 97 40 50 53 53 44 32 50 53 53 44 32 50 53 53 44 32 48 41 59 34 62 8235 98 33 8236 60 47 115 112 97 110 62 60 47 102 111 110 116 62
expected (character in Unicode value): 8235 98 33 8236
